fix: fix adding bot to group

+++ tg_bot/modules/sql/cust_filters_sql.py
@@ -1,6 +1,6 @@
 import threading
 
-from sqlalchemy import Column, String, UnicodeText, Boolean, Integer, distinct, func
+from sqlalchemy import Column, String, UnicodeText, Boolean, BigInteger, distinct, func
 
 from tg_bot.modules.sql import BASE, SESSION
 
@@ -46,7 +46,7 @@ class CustomFilters(BASE):
 
 class Buttons(BASE):
     __tablename__ = "cust_filter_urls"
-    id = Column(Integer, primary_key=True, autoincrement=True)
+    id = Column(BigInteger, primary_key=True, autoincrement=True)
     chat_id = Column(String(14), primary_key=True)
     keyword = Column(UnicodeText, primary_key=True)
     name = Column(UnicodeText, nullable=False)
