+++ tg_bot/modules/sql/users_sql.py
@@ -1,6 +1,6 @@
 import threading
 
-from sqlalchemy import Column, Integer, BigInteger, UnicodeText, String, ForeignKey, UniqueConstraint, func
+from sqlalchemy import Column, BigInteger, UnicodeText, String, ForeignKey, UniqueConstraint, func
 
 from tg_bot import dispatcher
 from tg_bot.modules.sql import BASE, SESSION
@@ -34,14 +34,14 @@ class Chats(BASE):
 
 class ChatMembers(BASE):
     __tablename__ = "chat_members"
-    priv_chat_id = Column(Integer, primary_key=True)
+    priv_chat_id = Column(BigInteger, primary_key=True)
     # NOTE: Use dual primary key instead of private primary key?
     chat = Column(String(14),
                   ForeignKey("chats.chat_id",
                              onupdate="CASCADE",
                              ondelete="CASCADE"),
                   nullable=False)
-    user = Column(Integer,
+    user = Column(BigInteger,
                   ForeignKey("users.user_id",
                              onupdate="CASCADE",
                              ondelete="CASCADE"),
